post

+++ app/Http/Controllers/TasksController.php
@@ -14,11 +14,18 @@ class TasksController extends Controller
      */
     public function index()
     {
-         $tasks = Task::all(); 
-         
-         return view('tasks.index', [
-              'tasks' => $tasks, 
-               ]); 
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            
+            $tasks= $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        } 
+        return view('dashboard', $data);
     }
 
     /**
@@ -104,11 +111,14 @@ class TasksController extends Controller
             'content' => 'required|max:255',
         ]);
         
-        $task = Task::findOrFail($id);
-        // メッセージを更新
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        if (\Auth::id() === $task->user_id) {
+        
+            $task = Task::findOrFail($id);
+            // メッセージを更新
+            $task->status = $request->status;
+            $task->content = $request->content;
+            $task->save();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -122,9 +132,11 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+        if (\Auth::id() === $task->user_id) {
+            $task = Task::findOrFail($id);
+            // メッセージを削除
+            $task->delete();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
